Added initial versions of ticket and init.
The basic execution flow is now handled by callbacks from within
argparse. We accomplish this by using set_defaults in subparsers to set
callback commands, then rely on the fact that only the main command
options, along with those of the selected subcommand, persist in the
finished args structure. Thus, only the set_defaults from the selected
subcommand remain, so only that func is defined.