Added initial descriptions to all commands.
I added initial descriptions to all commands, as well as replaced the
init strings with ones more appropriate. This stub now nicely shows you
which subcommand you are trying to run. Better, the help messages are
more meaningful, and now show the list of subcommands. I may also take
advantage of the aliasing commands to let subcommands have more than one
name, but we'll see about that.

The next step is to figure out the sqlite3 connection and get init to
prepare a not crappy environment for the other commands to work in.
After we have a working schema and an init that actually works, we
should be able to create the first actual tickets in the system!