Adding basic sketches of index and ticket.
I am adding some basic sketches of the index and ticket model classes.
Tickets are very simple, and I suspect that the class is more or less
done. The index still has several factory methods and a pile of
accessors to implement, but we are well on our way there. Once the index
interface is more settled, I'll go ahead and start implementing the
DatabaseModel class, which hews to the same interface.