Added some comments.
I thought about how some of the methods could work. I wrote these
thoughts as comments. When I can get the network back, I'll look up how
to serialize and de-serialize objects in Python. It should be pretty
simple. Once I have read and write object methods, the rest should be
pretty simple. We can determine if the index has a ticket by trying to
get that ticket's ID's object. At the same point, that's also how we can
try to reify it. Only if that fails, in some cases, will we then try to
use the database to reify.