Added code to read and write index.
I decided to use shelve object persistence to model the index. Thus, the
index model is vastly simplified. This commit contains the code to
actually read and write tickets. I believe that, unit tests aside, this
section of code is now complete. As for testing, unit tests will be done
in python, and system and acceptance tests will be handled directly by
ctest. This way, I can put all testing under the aegis of a single make
target.