Started sketching out the controller.
I tried to sketch out the basic tasks that the controller should be
responsible for. I believe that, for the most part, a simple one-to-one
relationship between sub commands and controller task methods is
appropriate, with the odd exception or two.

As for the basic workflow model, Ticket will have two places where
tickets can reside: the ticket database, and the ticket index. The
database is modelled by DatabaseModel, and is responsible for the
long-term storage and querying of ticket information. The ticket index
is responsible for the temporary storage of tickets that are
works-in-progress. Whenever a ticket's state is changed, the altered
ticket is stored in the index. Only when the commit sub command is run,
does the ticket return to the database for long-term storage.

The main advantage for developing the system this way is the fact that
the editing commands need only deal with the index, which promises to be
speedy, rather than the database, which promises to never forget.